requirements.txt
tweepy
openai
python-dotenv
quart
uvicorn